Partway through leveldb engine testing
